fix(native/text): generalize responsiveSize function to a utils file
